refactor(#792): split infra-filesystem lib_tests into recents + settings suites (sweep)
lib_tests.rs 712 -> 285, + lib_settings_tests.rs 435 (both under cap; off debt).
The GuiSystemSettings + GuiAudioDeviceSettings tests move to the sibling; tmp_dir
and make_device become pub(super), imported via super::tests. infra-filesystem
tests pass, zero warnings.
