+++ scripts/validate.sh
@@ -72,7 +72,6 @@ crates/infra-cpal/src/stream_builder.rs
 crates/infra-cpal/src/tests_regression.rs
 crates/infra-cpal/src/tests.rs
 crates/infra-cpal/tests/issue_670_cab_swap.rs
-crates/infra-filesystem/src/lib_tests.rs
 crates/infra-yaml/src/lib_tests.rs
 crates/lv2/src/host.rs
 crates/plugin-loader/src/manifest_tests.rs
